requirements.txt
# Core dependencies
pandas==2.3.0
requests==2.32.3
python-dotenv==1.1.0

# Web scraping
lxml==5.4.0
beautifulsoup4==4.13.4

# Vector database and embeddings
chromadb==1.0.12
openai==1.85.0
numpy==2.3.0

# LangChain ecosystem
langchain==0.3.25
langchain-openai==0.3.21
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langsmith==0.3.45

# Data validation
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2

# FastAPI dependencies
fastapi==0.115.9
uvicorn==0.34.3
starlette==0.45.3

# Streamlit dependencies
streamlit==1.45.1

# Testing
pytest==8.4.0

# Text processing and utilities
tiktoken==0.9.0
rank-bm25==0.2.2
typing_extensions==4.13.2

# Additional core libraries
PyYAML==6.0.2
SQLAlchemy==2.0.41
tenacity==9.1.2